security: hostile-input matrix, and the six allocations it found
The exit criterion for FREEZE.md's first security box: format/hostile_test.go
drives truncation at every byte offset of a complete body, every count at 0, 1,
max and max+1, and a set of allocation ceilings, through ReadWorld,
ReadStructure, ReadMeta and OpenIndexed.

The matrix is designed around STATUS.md's lesson rather than around truncation:
a count checked against the bytes that remain is not bounded if each unit of it
allocates more than a byte. Six places had that shape, and each fix caps a
capacity hint or moves an allocation behind a check the input has to pass
anyway. None of them changes which files a reader accepts; the golden and
vector suites were green throughout with no flags.

  blob table count       1,634 B -> 2.42 GiB, now 48 MiB (the body itself)
  block palette count      158 B ->   34 MiB, now 2.4 MiB
  biome palette count      159 B ->   63 MiB, now 2.5 MiB
  directory segment count  165 B ->   26 MiB, now 2.4 MiB
  structure cell grid       87 B ->    8 MiB, now 4.7 KiB
  verifyRecords          4,226 B ->   67 MiB, now 2.9 MiB

Also: UnresolvedStates read palette segments outside the lock, racing Compact,
which replaces the file handle, the compression state and the dictionary codec
-- and readFrame does not re-check a frame hash, so old offsets read against a
new file decoded whatever sat at them. And blobIndices compared a palette
length through a uint16; unreachable today, hardened anyway.

Two findings are reported and pinned rather than fixed, because closing either
rejects a file this reader accepts: the §8 NBT container budget does not charge
compounds nested in compounds, and §3.1's version-override index chain wraps in
uint64 before its bounds check. SECURITY.md says precisely which files each
would invalidate.

SECURITY.md also records the threat model FREEZE.md asks for -- xxHash64 is not
a cryptographic hash, so an attacker who can author content and induce
truncation can forge a checkpoint -- and what remains: the decompression ratio
against ceilings §8 sets high (a legal 1,161-byte file decodes into 1.04 GiB),
and recovery's 256 candidates times a 4M-entry directory, which is fourteen
minutes from a 20 KB file. Bounding either is a format change.

Ticks the three security boxes this covers plus the threat model. Each fix has
a negative control recorded in SECURITY.md.

+++ format/decode.go
@@ -832,23 +832,29 @@ func decodeRecordBody(r *reader, reg world.BlockRegistry, rids, biomeIDs []uint3
 // blobIndices decodes a blob's 4096 local indices into out, validating them
 // against the local palette size.
 func blobIndices(blob decBlob, paletteLen int, out *[4096]uint16) error {
-	n := uint16(paletteLen)
+	// Compared as an int, not through a uint16: the palette count a blob may
+	// declare reaches 65,536, which narrows to zero and would then reject
+	// every index in the blob that declared it. Nothing reaches that today,
+	// because §3.3 requires every declared entry to be named by one of 4096
+	// indices and so refuses any palette larger than 4096 first. The bound is
+	// written not to depend on that argument holding somewhere else.
+	n := paletteLen
 	switch blob.width {
 	case widthU8:
 		for i := range 4096 {
-			li := uint16(blob.idx[i])
+			li := int(blob.idx[i])
 			if li >= n {
 				return corruptf("section index %d out of palette range %d", li, n)
 			}
-			out[i] = li
+			out[i] = uint16(li)
 		}
 	case widthU16:
 		for i := range 4096 {
-			li := uint16(blob.idx[i*2]) | uint16(blob.idx[i*2+1])<<8
+			li := int(blob.idx[i*2]) | int(blob.idx[i*2+1])<<8
 			if li >= n {
 				return corruptf("section index %d out of palette range %d", li, n)
 			}
-			out[i] = li
+			out[i] = uint16(li)
 		}
 	}
 	return nil
